Update tfidf.py
Updated class TfIdfIndex to handle hash collisions that were possible with the _hash32 function. 
added self._id_to_i dictionary to hold a reverse pairing of key/values used in self._i_to_id.
updated _hash32 to check for the hash value in the dictionary keys before adding and updating the hash if necessary.
updated the index function to assign values to both dictionaries.
updated unindex to pull the value for i from the _id_to_i dictionary to ensure it is getting a record_id that exists.

Let me know if you have any questions.