Add a join method to MockPool
A call to `pool.join()` was added in a previous commit to fix a problem with
spawned subprocesses launched by `StaticGazetteer.match` never exiting.

When running with a single core, `pool` is a `MockPool` instance rather than a
`multiprocessing.Pool`. The `MockPool` needs to have a `join` method to avoid
an unhandled `AttributeError`.