Sample should be sample w/replacement not w/o replacement
for bigger data (over 5K records) we draw a random sample for purpose
of estimating the comparison counts of predicates. If we sample w/ replacement
then we have a chance of pulling in exact duplicates, which can cause
us to overestimate the number of comparisons. this is especially
noticeable for precidates that have low comparison count