Make use of Index and Canopy in docs consistent
I don't really understand the difference between Index
and Canopy blocks (Index blocks are a type of blocks,
and canopy clustering is an algorithm to implement them?),
but I was getting a bit confused from reading the docs as to
the difference between them, and how they relate to the
`index_predicates` argument to the `train()` meethods.

So, this is my best guess as to how to make their use in
the docs more consistent.