Fix docs build errors
When GH actions ran sphinx, I got the error:
Warning, treated as error:
/home/runner/work/dedupe/dedupe/docs/API-documentation.rst:138:duplicate object description of dedupe.Dedupe, other instance in API-documentation, use :noindex: for one of them

So I fixed that. Didn't seem like it changed anything.

But, there was a real bug where `Fingerprinter.index_fields`
wasn't getting any documentation generated for it.
(check out the docs that are currently live) So fix that.