Split examples into their own doc page
Makes it easier to find, since they are now just always
on the sidebar. I think this is a common thing people
are looking for.

Also, fix the formatting of this section. Make it a bit clearer the
two types of examples there are. And, remove the link to the
.zip file of examples. It is destined to get out of date, and
people can download the most up-to-date version from the repo.