Make benchmarks installable
Not sure this is the best way to do this, but
I was running intor problems because asv is run inside it's own
little venv, so you can't just import neighboring modules by name
because you don't know the path that the interpreter starts on.

I'm pretty confused by all this though, so I'm mostly just thrashing
around until I find somethign that works.