Move all_subclasses() to classmethod
It assumes that the `type` attribute exists, so really it should
be associated with the Variable class.
Where this is used in datamodel, we use it on FieldVariable, so any
plugins that want to be registered with dedupe already have to subclass
from FieldVariable, so this doesn't break anything.

I tweaked it to use getattr()
so that any abstract base classes
can just omit setting `type` instead of
having to explicitly set it.
This does have the effect in plugins where if you forget to set `type`
in your custom subclass, the subclass doesn't get registered.
previously it would throw an error. But as soon as that dev tries to
use their custom subclass, it's not gonna be found, so should
be pretty obvious.