Use conventions for fit(), fit_transform(), etc.
The meaning of these methods is not consistent with how sklearn uses
them.

For them, fit_transform(X, y) means
fit(X, y).transform(X), but we were using it as
fit(transform(X), y). So just rename fit_transform() to fit(), and now
it has the correct meaning.

What's more, the transform method isn't used external to the class, and
I don't think it should be. The public API should only deal with
TrainingPairs, it shouldn't deal with the
calculated distances. Rename it to _distances()

The old fit() method now is just an internal helper method that deals
with already calculated distance data.