Delegate DisagreementLearner.learn_predicates()
DisagreementLearner is really reaching far down into the contained
objects. The lower classes themselves should be responsible for this
sort of thing. Sure, it makes the code more complicated, but the way it
is is fooling ourselves that it is simple,
and making it prone to breaking int he future.