Ensure unique variable names
Before, we never checked for this,
so when we called list.index() we
just got back the first instance.

I swapped us over to a lookup table because that makes more
sense, but I wanted to also add this check because if someone had
duplicate names, this would quitely change the behavior underneath
someone: before it gave the first index, now it gives the last.