ref: Create InteractionVariables the same as the rest
Before the flow was:
1. create all the primary variables EXCEPT for the interactions
2. Expand those.
3. go back through and NOW create the InteractionVariables

I found this confusing. We parse the var definitions twice in separate
places, and InteractionVariables
are a special case in the first pass.

The other point of this is to make it so that there
is one list of Variable instances
which is the single source of truth. Once we do this
then we can turn all the other private instance variables of DataModel
into @functools.cached_property's,
which will make it much easier to
ensure pickle compatibility in
the future, because
only one variable needs to get
saved and restored